Second Commit
In this project,Every post has title , author, content, datetime. Guest
can read a post and see details of it, make a comment.
Admin can create a post , delete a post, update a post,show details of
it, read a post ,make a comment to the post and add a file.